refactor: Pass diagram as React.Node
