--- labware-library/src/components/MeasurementGuide/styles.css
@@ -1,15 +1,0 @@
-@import '@opentrons/components';
-@import '../../styles/breakpoints.css';
-@import '../../styles/spacing.css';
-
-.measurement_guide {
-  height: auto;
-  max-height: 0;
-  overflow: hidden;
-  transition: max-height 0.5s linear;
-
-  &.open {
-    /* TODO: investigate react handling dynamic max height */
-    max-height: 800px;
-  }
-}
